+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/MyRobot.java
@@ -8,7 +8,7 @@ import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.teamcode.subSystems.Bucket;
 import org.firstinspires.ftc.teamcode.subSystems.Wrist;
 import org.firstinspires.ftc.teamcode.subSystems.Xarm;
-import org.firstinspires.ftc.teamcode.subSystems.Yarm;
+import org.firstinspires.ftc.teamcode.subSystems.YArm;
 
 /**
  * MyRobot class that extends RobotContext to include robot-specific subsystems.
@@ -16,7 +16,7 @@ import org.firstinspires.ftc.teamcode.subSystems.Yarm;
 public class MyRobot extends RobotContext {
     public final MecanumDrive DRIVE;
     public final Xarm XARM;
-    public final Yarm YARM;
+    public final YArm YARM;
     public final Wrist WRIST;
     public final Bucket BUCKET;
 
@@ -31,11 +31,11 @@ public class MyRobot extends RobotContext {
      * @param gamepad1  the primary gamepad controller
      * @param gamepad2  the secondary gamepad controller
      */
-    public MyRobot(Telemetry telemetry, Gamepad gamepad1, Gamepad gamepad2, MecanumDrive drive, Xarm xarm, Yarm yarm, Wrist wrist, Bucket bucket, Yarm yarm1, Wrist wrist1, Bucket bucket1) {
+    public MyRobot(Telemetry telemetry, Gamepad gamepad1, Gamepad gamepad2, MecanumDrive drive, Xarm xarm, YArm yarm, Wrist wrist, Bucket bucket, YArm YArm1, Wrist wrist1, Bucket bucket1) {
         super(telemetry, gamepad1, gamepad2);
         this.DRIVE = drive;
         this.XARM = xarm;
-        this.YARM = yarm1;
+        this.YARM = YArm1;
         this.WRIST = wrist1;
         this.BUCKET = bucket1;
